solve bug when clicking on the selectedOption
