+++ scripts/utils/sortButton.js
@@ -1,0 +1,52 @@
+export function openOptionsList() {
+  const button = document.querySelector(".sort__button");
+  const list = document.querySelector(".sort__options");
+  button.addEventListener("click", function() {
+    const sort = document.querySelector(".sort__select");
+    sort.classList.toggle("sort__select-open");
+    if (sort.classList.contains("sort__select-open")) {
+      button.setAttribute("aria-expanded", "true");
+      sort.querySelector("i").classList.toggle("fa-caret-up");
+      sort.querySelector("i").classList.remove("fa-caret-down");
+      list.style.display = "block";
+    } else {
+      button.setAttribute("aria-expanded", "false");
+      sort.querySelector("i").classList.toggle("fa-caret-down");
+      sort.querySelector("i").classList.remove("fa-caret-up");
+      list.style.display = "none";
+    }
+  })
+}
+function closeOptionsList() {
+  const button = document.querySelector(".sort__button");
+  const list = document.querySelector(".sort__options");
+  const sort = document.querySelector(".sort__select");
+  button.setAttribute("aria-expanded", "false");
+  sort.querySelector("i").classList.add("fa-caret-down");
+  sort.querySelector("i").classList.remove("fa-caret-up");
+  sort.classList.toggle("sort__select-open");
+  list.style.display = "none"
+}
+
+export function selectOption() {
+  const options = document.querySelectorAll(".sort__option");
+  const orderBtn = document.querySelector(".sort__button");
+
+  options.forEach(option => {
+    option.classList.remove("sort__hide");
+    option.addEventListener("click", (event) => {
+      orderBtn.innerText = event.target.innerText;
+      option.classList.add("sort__hide");
+      option.setAttribute("aria-selected", "true");
+      closeOptionsList();
+      const selectedOption = Array.from(options).find(option => option.innerText === orderBtn.innerText);
+      selectedOption.style.display = "none"
+      const otherOptions = Array.from(options).filter(option => option.innerText !== orderBtn.innerText);
+      otherOptions.forEach(option => function() {
+        option.style.display = "block";
+        option.setAttribute("aria-selected", "false");
+        option.style.display = "block"
+      })
+    })
+  })
+}
